add real skip list insert
